lsp: Move node ignore code into the common section
... it is conceptually meant to be used by the language server and the
live-preview.
